Extract dialogue implementations from skills and add to fipa protocol
