+++ packages/skills/tac_negotiation/helpers.py
@@ -21,16 +21,12 @@
 """This class contains the helpers for FIPA negotiation."""
 
 import copy
-from typing import Dict, List, Union, TYPE_CHECKING, cast
+from typing import Dict, List, Union, cast
 
 from aea.decision_maker.messages.transaction import TransactionMessage
+from aea.helpers.dialogue.base import DialogueLabel
 from aea.protocols.oef.models import Attribute, DataModel, Description, Query, Constraint, ConstraintType, Or, \
     ConstraintExpr
-
-if TYPE_CHECKING:
-    from packages.skills.tac_negotiation.dialogues import DialogueLabel
-else:
-    from tac_negotiation_skill.dialogues import DialogueLabel
 
 Address = str
 TransactionId = str
